add script

+++ env/env-msys2/utils-git.sh
@@ -41,25 +41,27 @@ _utils_git_init() {
 	alias gcv='git commit --no-verify'
 }
 
-utils_git_set() {	
+utils_git_set() {
 	template_prefix=$QDKE_ROOT/env/env-msys2/tpl-githooks
 	dst_hooks_dir=$work_home/.git/hooks
-
-	# exe_cmd "mkdir -p $template_dir"
-	exe_cmd "cp -rf $template_prefix-pre-commit.sh $dst_hooks_dir/pre-commit.sh"
-	exe_cmd "chmod -R a+x $dst_hooks_dir/pre-commit.sh"
-	# exe_cmd "git config --global init.templatedir $template_dir"
-	# exe_cmd "git config --global core.fileMode false"
-	git config --global alias.st 'status'
-	git config --global alias.ci 'commit'
-	git config --global alias.sb 'submodule'
-	git config --global alias.co 'checkout'
-	git config --global alias.pm 'push origin master'
-	git config --global alias.pd 'push origin dev'
-	git config --global alias.lg "log --all --graph --pretty=format:'%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%ci) %C(bold blue)<%an>%Creset' --abbrev-commit"
-	git config --global alias.d difftool
-	git config --global diff.tool vimdiff
-	git config --global difftool.prompt false
+	
+	if [ ! -f $dst_hooks_dir/pre-commit.sh ]; then
+		# exe_cmd "mkdir -p $template_dir"
+		exe_cmd "cp -rf $template_prefix-pre-commit.sh $dst_hooks_dir/pre-commit.sh"
+		exe_cmd "chmod -R a+x $dst_hooks_dir/pre-commit.sh"
+		# exe_cmd "git config --global init.templatedir $template_dir"
+		# exe_cmd "git config --global core.fileMode false"
+		git config --global alias.st 'status'
+		git config --global alias.ci 'commit'
+		git config --global alias.sb 'submodule'
+		git config --global alias.co 'checkout'
+		git config --global alias.pm 'push origin master'
+		git config --global alias.pd 'push origin dev'
+		git config --global alias.lg "log --all --graph --pretty=format:'%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%ci) %C(bold blue)<%an>%Creset' --abbrev-commit"
+		git config --global alias.d difftool
+		git config --global diff.tool vimdiff
+		git config --global difftool.prompt false
+	fi
 }
 
 utils_git_check() {
